backend/requirements.txt

fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
pyjwt
python-multipart
python-dotenv
itsdangerous
